requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
ultralytics 
pillow   
numpy
python-dotenv
pdf2image
pytesseract
pymysql 
pymupdf 
google-genai
scikit-learn
joblib
torch
torchvision
